+++ diesel/src/query_builder/insert_statement/mod.rs
@@ -15,7 +15,7 @@ use super::returning_clause::*;
 use crate::backend::{DieselReserveSpecialization, SqlDialect, sql_dialect};
 use crate::expression::grouped::Grouped;
 use crate::expression::operators::Eq;
-use crate::expression::{Expression, NonAggregate, SelectableExpression};
+use crate::expression::{Expression, NonAggregate};
 use crate::query_builder::*;
 use crate::query_dsl::RunQueryDsl;
 use crate::query_source::{Column, Table};
@@ -286,6 +286,7 @@ impl<T, U, Op> AsQuery for InsertStatement<T, U, Op, NoReturningClause>
 where
     T: Table,
     InsertStatement<T, U, Op, ReturningClause<T::AllColumns>>: Query,
+    T::AllColumns: ReturningExpression<InsertStmt, T>,
 {
     type SqlType = <Self::Query as Query>::SqlType;
     type Query = InsertStatement<T, U, Op, ReturningClause<T::AllColumns>>;
@@ -298,9 +299,9 @@ where
 impl<T, U, Op, Ret> Query for InsertStatement<T, U, Op, ReturningClause<Ret>>
 where
     T: QuerySource,
-    Ret: Expression + SelectableExpression<T> + NonAggregate,
+    Ret: ReturningExpression<InsertStmt, T> + NonAggregate,
 {
-    type SqlType = Ret::SqlType;
+    type SqlType = <Ret as ReturningExpression<InsertStmt, T>>::SqlType;
 }
 
 impl<T: QuerySource, U, Op, Ret, Conn> RunQueryDsl<Conn> for InsertStatement<T, U, Op, Ret> {}